Extract e2e mock app settings helpers
